Error reporting, failure handling
Should send messages back to game when connections fail, so it can reset the UI instead of just stalling where it is.
It attempts to send fail message to server if hosting (to let other clients know what went wrong) or to checkout client if peer. Then it disconnects sockets and stops procceses.

Time for a lot of testing